Extract a Publish module
This keeps the code organized into related functionality
