+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2020-08-10 16:06:50 UTC using RuboCop version 0.89.0.
+# on 2020-08-18 22:05:35 UTC using RuboCop version 0.89.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -38,7 +38,7 @@ Metrics/ClassLength:
 Metrics/CyclomaticComplexity:
   Max: 12
 
-# Offense count: 62
+# Offense count: 61
 # Configuration parameters: CountComments, CountAsOne, ExcludedMethods.
 Metrics/MethodLength:
   Max: 43
@@ -66,20 +66,20 @@ Naming/HeredocDelimiterNaming:
 # SupportedStyles: snake_case, camelCase
 Naming/VariableName:
   Exclude:
-    - 'app/services/public_desc_metadata_service.rb'
+    - 'app/services/publish/public_desc_metadata_service.rb'
 
 # Offense count: 17
 RSpec/AnyInstance:
   Exclude:
     - 'spec/requests/create_object_spec.rb'
     - 'spec/requests/notify_goobi_spec.rb'
-    - 'spec/services/public_xml_service_spec.rb'
-    - 'spec/services/publish_metadata_service_spec.rb'
+    - 'spec/services/publish/metadata_transfer_service_spec.rb'
+    - 'spec/services/publish/public_xml_service_spec.rb'
 
 # Offense count: 21
 RSpec/Be:
   Exclude:
-    - 'spec/services/public_xml_service_spec.rb'
+    - 'spec/services/publish/public_xml_service_spec.rb'
 
 # Offense count: 2
 RSpec/BeforeAfterAll:
@@ -96,9 +96,9 @@ RSpec/ContextWording:
   Exclude:
     - 'spec/dor/update_marc_record_service_spec.rb'
     - 'spec/services/digital_stacks_service_spec.rb'
-    - 'spec/services/public_desc_metadata_service_spec.rb'
-    - 'spec/services/public_xml_service_spec.rb'
-    - 'spec/services/publish_metadata_service_spec.rb'
+    - 'spec/services/publish/metadata_transfer_service_spec.rb'
+    - 'spec/services/publish/public_desc_metadata_service_spec.rb'
+    - 'spec/services/publish/public_xml_service_spec.rb'
     - 'spec/services/release_tags/identity_metadata_spec.rb'
     - 'spec/services/thumbnail_service_spec.rb'
 
@@ -121,8 +121,8 @@ RSpec/ExpectInHook:
   Exclude:
     - 'spec/dor/update_marc_record_service_spec.rb'
     - 'spec/services/datastream_extractor_spec.rb'
-    - 'spec/services/public_desc_metadata_service_spec.rb'
-    - 'spec/services/publish_metadata_service_spec.rb'
+    - 'spec/services/publish/metadata_transfer_service_spec.rb'
+    - 'spec/services/publish/public_desc_metadata_service_spec.rb'
     - 'spec/services/sdr_ingest_service_spec.rb'
     - 'spec/services/shelving_service_spec.rb'
 
@@ -140,7 +140,7 @@ RSpec/MessageSpies:
     - 'spec/dor/update_marc_record_service_spec.rb'
     - 'spec/services/datastream_extractor_spec.rb'
     - 'spec/services/digital_stacks_service_spec.rb'
-    - 'spec/services/public_desc_metadata_service_spec.rb'
+    - 'spec/services/publish/public_desc_metadata_service_spec.rb'
     - 'spec/services/sdr_ingest_service_spec.rb'
     - 'spec/services/shelving_service_spec.rb'
     - 'spec/services/version_service_spec.rb'
@@ -167,15 +167,15 @@ RSpec/PredicateMatcher:
 # Offense count: 3
 RSpec/ScatteredLet:
   Exclude:
-    - 'spec/services/public_desc_metadata_service_spec.rb'
-    - 'spec/services/public_xml_service_spec.rb'
+    - 'spec/services/publish/public_desc_metadata_service_spec.rb'
+    - 'spec/services/publish/public_xml_service_spec.rb'
     - 'spec/services/shelving_service_spec.rb'
 
 # Offense count: 4
 RSpec/ScatteredSetup:
   Exclude:
-    - 'spec/services/public_desc_metadata_service_spec.rb'
-    - 'spec/services/publish_metadata_service_spec.rb'
+    - 'spec/services/publish/metadata_transfer_service_spec.rb'
+    - 'spec/services/publish/public_desc_metadata_service_spec.rb'
 
 # Offense count: 3
 RSpec/SubjectStub:
@@ -217,7 +217,7 @@ Style/ConditionalAssignment:
   Exclude:
     - 'spec/support/foxml_helper.rb'
 
-# Offense count: 11
+# Offense count: 12
 Style/Documentation:
   Exclude:
     - 'spec/**/*'
@@ -226,8 +226,8 @@ Style/Documentation:
     - 'app/controllers/objects_controller.rb'
     - 'app/controllers/versions_controller.rb'
     - 'app/models/dor/update_marc_record_service.rb'
-    - 'app/services/dublin_core_service.rb'
     - 'app/services/metadata_service.rb'
+    - 'app/services/publish/dublin_core_service.rb'
     - 'app/services/release_tags/identity_metadata.rb'
     - 'config/application.rb'
     - 'config/initializers/okcomputer.rb'
@@ -294,19 +294,18 @@ Style/PercentLiteralDelimiters:
     - 'config/spring.rb'
     - 'spec/dor/update_marc_record_service_spec.rb'
 
-# Offense count: 24
+# Offense count: 23
 # Cop supports --auto-correct.
 Style/StringConcatenation:
   Exclude:
-    - 'app/controllers/content_controller.rb'
-    - 'app/services/dublin_core_service.rb'
-    - 'app/services/public_xml_service.rb'
+    - 'app/services/publish/dublin_core_service.rb'
+    - 'app/services/publish/public_xml_service.rb'
     - 'bin/console'
     - 'spec/rails_helper.rb'
     - 'spec/requests/create_workspaces_spec.rb'
     - 'spec/services/cleanup_service_spec.rb'
-    - 'spec/services/public_desc_metadata_service_spec.rb'
-    - 'spec/services/public_xml_service_spec.rb'
+    - 'spec/services/publish/public_desc_metadata_service_spec.rb'
+    - 'spec/services/publish/public_xml_service_spec.rb'
     - 'spec/services/reset_workspace_service_spec.rb'
 
 # Offense count: 15
@@ -336,7 +335,7 @@ Style/ZeroLengthPredicate:
   Exclude:
     - 'app/models/dor/update_marc_record_service.rb'
 
-# Offense count: 290
+# Offense count: 291
 # Cop supports --auto-correct.
 # Configuration parameters: AutoCorrect, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
 # URISchemes: http, https
