+++ spec/jobs/publish_job_spec.rb
@@ -25,7 +25,7 @@ RSpec.describe PublishJob, type: :job do
 
   context 'with no errors' do
     before do
-      allow(PublishMetadataService).to receive(:publish)
+      allow(Publish::MetadataTransferService).to receive(:publish)
       allow(LogSuccessJob).to receive(:perform_later)
       perform
     end
@@ -34,8 +34,8 @@ RSpec.describe PublishJob, type: :job do
       expect(result).to have_received(:processing!).once
     end
 
-    it 'invokes the PublishMetadataService' do
-      expect(PublishMetadataService).to have_received(:publish).with(item).once
+    it 'invokes the Publish::MetadataTransferService' do
+      expect(Publish::MetadataTransferService).to have_received(:publish).with(item).once
     end
 
     it 'marks the job as complete' do
@@ -46,11 +46,11 @@ RSpec.describe PublishJob, type: :job do
     end
   end
 
-  context 'with errors returned by PublishMetadataService' do
+  context 'with errors returned by Publish::MetadataTransferService' do
     let(:error_message) { "DublinCoreService#ng_xml produced incorrect xml (no children):\n<xml/>" }
 
     before do
-      allow(PublishMetadataService).to receive(:publish).and_raise(Dor::DataError, error_message)
+      allow(Publish::MetadataTransferService).to receive(:publish).and_raise(Dor::DataError, error_message)
       allow(LogFailureJob).to receive(:perform_later)
       perform
     end
@@ -59,8 +59,8 @@ RSpec.describe PublishJob, type: :job do
       expect(result).to have_received(:processing!).once
     end
 
-    it 'invokes the PublishMetadataService' do
-      expect(PublishMetadataService).to have_received(:publish).with(item).once
+    it 'invokes the Publish::MetadataTransferService' do
+      expect(Publish::MetadataTransferService).to have_received(:publish).with(item).once
     end
 
     it 'marks the job as complete' do
